refactor(DisplayController.js): make class to file
