+++ src/App.js
@@ -1,29 +1,8 @@
 import { Console } from "@woowacourse/mission-utils";
-import RACING_VARIABLES from './constants/RacingVariables';
 import RaceController from "./controllers/RaceController";
 import Validator from "./utils/Validator";
 import UserInput from "./views/UserInput";
-
-
-
-
-
-class DisplayController{
-
-  async displayPlayerPositions (players, positions){
-    for (let i = 0; i<players.length;i++){
-      Console.print(`${players[i]} : ${ RACING_VARIABLES.POSITION_MARKER.repeat(positions[i])}`);
-    }
-    Console.print("");
-  }
-
-  async displayWinners (players, winners){
-    const winnerArray = winners.map(idx => players[idx])
-
-    Console.print(`${RACING_VARIABLES.WINNER_PROMPT} : ${winnerArray.join(',')}`)
-  }
-}
-
+import DisplayController from "./views/DisplayController";
 
 class App {
   constructor (){
